requirements.txt
blinker==1.6.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.6
Flask==3.0.0
gunicorn==21.2.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==2.1.2
packaging==23.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.0
